+++ src/aws/cis-1.2.0/rules/aws-cis-1.2.0-1.8.ts
@@ -1,0 +1,18 @@
+export default {
+  id: 'aws-cis-1.2.0-1.8',
+  description:
+    'AWS CIS 1.8  Ensure IAM password policy requires at least one number',
+  gql: `{
+    queryawsIamPasswordPolicy {
+      id
+      __typename
+      requireNumbers
+    }
+  }`,
+  resource: 'queryawsIamPasswordPolicy[*]',
+  severity: 'warning',
+  conditions: {
+    path: '@.requireNumbers',
+    equal: true,
+  },
+}