+++ src/aws/cis-1.2.0/rules/aws-cis-1.2.0-2.9.ts
@@ -1,0 +1,20 @@
+export default {
+  id: 'aws-cis-1.2.0-2.9',
+  description:
+    'AWS CIS 2.9 Ensure VPC flow logging is enabled in all VPCs (Scored)',
+  gql: `{
+    queryawsVpc {
+      id
+      __typename
+      flowLogs {
+        resourceId
+      }
+    }
+  }`,
+  resource: 'queryawsVpc[*]',
+  severity: 'warning',
+  conditions: {
+    path: '@.flowLogs',
+    isEmpty: false,
+  },
+}