+++ src/aws/cis-1.3.0/rules/aws-cis-1.3.0-1.5.ts
@@ -1,0 +1,18 @@
+export default {
+  id: 'aws-cis-1.3.0-1.5',
+  description:
+    'AWS CIS 1.5  Ensure IAM password policy requires at least one uppercase letter',
+  gql: `{
+    queryawsIamPasswordPolicy {
+      id
+      __typename
+      requireUppercaseCharacters
+    }
+  }`,
+  resource: 'queryawsIamPasswordPolicy[*]',
+  severity: 'warning',
+  conditions: {
+    path: '@.requireUppercaseCharacters',
+    equal: true,
+  },
+}