+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.1.0](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/compare/1.0.0...1.1.0) (2021-11-17)
+
+
+### Bug Fixes
+
+* Added tests to cis 1.10 rule ([fa05117](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/fa05117b0b81100c0a72d9d50bb0d92d95d36869))
+* Updated rule cis 1.2 to use isEmpty operator ([c807068](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/c80706836f1fd2e33ab165c2ef11cb0309120d59))
+
+
+### Features
+
+* Added password policy requires minimum length of 14 or greater rule ([8774124](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/8774124af243267ba75fbdae80eb2652d86a8a5e))
+* Added rule to prevent IAM password reuse ([6e6c45e](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/6e6c45e8f1b33fe02af48f472ea3e65514efffa1))
+* Enabled unit tests for existing rules ([77304d5](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/77304d57f48896cded26a55185e0cda47b5c9a48))
+* Included cis 1.1 rule ([b32c87c](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/b32c87cc68ad4d31f2efe589d5ded548233d2ef0))
+* Included cis 1.11 rule ([9ff45f2](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/9ff45f232d8e12c628c4e43339cf8e5be22e0dbf))
+* Included cis 1.12 rule ([0e331f3](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/0e331f35be6e5c052c9f4e314d9ec54f5a178fb7))
+* Included cis 1.13 rule ([134b504](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/134b5042eff219bc6fda71db9ed520aec1181c42))
+* Included cis 1.14 rule ([cafae2f](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/cafae2fe9c97f01e3943a2c842e8a1cf5ab54686))
+* Included cis 1.16 rule ([1b38c7a](https://gitlab.com/auto-cloud/cloudgraph/policy-packs/commit/1b38c7a467e775f7bd9cdf470d3d60c271e96a6c))
+
 # 1.0.0 (2021-11-04)